cilium: split cgroups handling into own package
Split it off from sockops and move into its own package since it's
going to be used outside of sockops in subsequent commits.
